chore(typography-util): rename font files
